relocate files and folder
