+++ controllers/project.js
@@ -1,4 +1,4 @@
-const db = require("../../db/index");
+const db = require("../db/index");
 
 function createProject(body) {
     console.log("Creating...");
