+++ controllers/status.js
@@ -1,4 +1,4 @@
-const db = require("../../db/index");
+const db = require("../db/index");
 
 function createStatus(body) {
     console.log("function run");
